+++ src/components/domain/auth/login/LoginButton.tsx
@@ -1,0 +1,14 @@
+import LoginModal from './LoginModal';
+import { useModalStore } from '@/stores/modalStore';
+
+interface LoginButtonProps {
+  children: React.ReactNode;
+}
+export default function LoginButton({ children }: LoginButtonProps) {
+  const { openModal } = useModalStore();
+
+  const handleOpenLoginModal = () => {
+    openModal(<LoginModal />);
+  };
+  return <div onClick={handleOpenLoginModal}>{children}</div>;
+}
